+++ controllers/hvpa_util.go
@@ -30,149 +30,6 @@ import (
 	vpa_api "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/apis/autoscaling.k8s.io/v1beta2"
 )
 
-const (
-	// ResourceCPU represents CPU in millicores (1core = 1000millicores).
-	ResourceCPU utils.AxisName = "cpu"
-	// ResourceMemory represents memory, in bytes. (500Gi = 500GiB = 500 * 1024 * 1024 * 1024).
-	ResourceMemory utils.AxisName = "memory"
-	// ResourceReplicas represents replicas.
-	ResourceReplicas utils.AxisName = "replicas"
-)
-
-func getLinearBucket(scaleInterval *hvpav1alpha1.EffectiveScalingInterval) (utils.Bucket, error) {
-	linearBucket := utils.NewLinearBucket()
-
-	maxReplicas := scaleInterval.Replicas
-	minReplicas := scaleInterval.Replicas
-	err := linearBucket.AddAxis(ResourceReplicas, int64(minReplicas), int64(maxReplicas))
-	if err != nil {
-		return nil, err
-	}
-
-	maxCPU := scaleInterval.MaxResources[hvpav1alpha1.ResourceCPU]
-	minCPU := scaleInterval.MinResources[hvpav1alpha1.ResourceCPU]
-	err = linearBucket.AddAxis(ResourceCPU, minCPU, maxCPU)
-	if err != nil {
-		return nil, err
-	}
-
-	maxMemory := scaleInterval.MaxResources[hvpav1alpha1.ResourceMemory]
-	minMemory := scaleInterval.MinResources[hvpav1alpha1.ResourceMemory]
-	err = linearBucket.AddAxis(ResourceMemory, minMemory, maxMemory)
-	if err != nil {
-		return nil, err
-	}
-
-	return linearBucket, nil
-}
-
-// Returns minAllowed of all containers for which VPA mode is not set to "Off"
-func getMinAllowed(containerPolicies []vpa_api.ContainerResourcePolicy) (minCPU, minMemory int64) {
-	var cpu, memory int64
-
-	for _, policy := range containerPolicies {
-		if policy.MinAllowed != nil && (policy.Mode == nil || *policy.Mode != vpa_api.ContainerScalingModeOff) {
-			if policy.MinAllowed.Cpu() != nil {
-				cpu += policy.MinAllowed.Cpu().MilliValue()
-			}
-			if policy.MinAllowed.Cpu() != nil {
-				memory += policy.MinAllowed.Memory().MilliValue()
-			}
-		}
-	}
-
-	return cpu, memory
-}
-
-func getBuckets(hvpa *hvpav1alpha1.Hvpa, currentReplicas int32) (bucketList, currentBucket utils.Bucket, err error) {
-	var (
-		buckets    []utils.Bucket
-		bucket     utils.Bucket
-		currBucket utils.Bucket
-		currMin    *int64
-		replicas   int32 = 1
-	)
-	scalingOverlap := hvpa.Spec.ScalingIntervalsOverlap
-
-	if hvpa.Spec.Hpa.Template.Spec.MinReplicas != nil {
-		replicas = *hvpa.Spec.Hpa.Template.Spec.MinReplicas
-	}
-	minCPU, minMemory := getMinAllowed(hvpa.Spec.Vpa.Template.Spec.ResourcePolicy.ContainerPolicies)
-
-	for _, scaleInterval := range hvpa.Spec.ScaleIntervals {
-		intervalMaxCPU := resource.MustParse(scaleInterval.MaxCPU)
-		intervalMaxMemory := resource.MustParse(scaleInterval.MaxMemory)
-		intervalMinCPU := minCPU
-		intervalMinMemory := minMemory
-
-		numOfSubIntervals := int64(scaleInterval.MaxReplicas - replicas + 1)
-		cpuDelta := (intervalMaxCPU.MilliValue() - intervalMinCPU) / numOfSubIntervals
-		memDelta := (intervalMaxMemory.MilliValue() - intervalMinMemory) / numOfSubIntervals
-
-		for i := int64(1); i <= numOfSubIntervals; i++ {
-			localMaxCPU := intervalMinCPU + i*cpuDelta
-			localMaxMem := intervalMinMemory + i*memDelta
-			subInterval := hvpav1alpha1.EffectiveScalingInterval{
-				Replicas: replicas,
-				MinResources: hvpav1alpha1.ResourceList{
-					hvpav1alpha1.ResourceCPU:    minCPU,
-					hvpav1alpha1.ResourceMemory: minMemory,
-				},
-				MaxResources: hvpav1alpha1.ResourceList{
-					hvpav1alpha1.ResourceCPU:    localMaxCPU,
-					hvpav1alpha1.ResourceMemory: localMaxMem,
-				},
-			}
-
-			bucket, err = getLinearBucket(&subInterval)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			if currentReplicas == replicas {
-				currBucket = bucket
-			}
-
-			log.V(4).Info("hvpa intervals", "interval", bucket.GetIntervals(), "hvpa", hvpa.Namespace+"/"+hvpa.Name)
-			buckets = append(buckets, bucket)
-
-			// Prepare for next iteration - total max of current bucket is equal to total min of next bucket
-			prevReplicas := replicas
-			replicas = replicas + 1
-			minCPU = localMaxCPU * int64(prevReplicas) / int64(replicas)
-			minMemory = localMaxMem * int64(prevReplicas) / int64(replicas)
-
-			if scalingOverlap != nil {
-				// Adjust min values for overlap
-				for resourceName, params := range scalingOverlap {
-					deltaPerc := int64(0)
-					deltaVal := int64(0)
-					if resourceName == corev1.ResourceCPU {
-						currMin = &minCPU
-					} else if resourceName == corev1.ResourceMemory {
-						currMin = &minMemory
-					} else {
-						log.V(3).Info("WARNING:", "unsupported resource in ScalingIntervalsOverlap", resourceName, "hvpa", hvpa.Namespace+"/"+hvpa.Name)
-						continue
-					}
-					perc := params.Percentage
-					if perc != nil {
-						deltaPerc = *currMin * int64(*perc) / 100
-					}
-					if params.Value != nil {
-						val := resource.MustParse(*params.Value)
-						deltaVal = val.MilliValue()
-					}
-					// Following statement will update minCPU/minMemory value
-					*currMin = *currMin - int64(math.Max(float64(deltaPerc), float64(deltaVal)))
-				}
-			}
-		}
-	}
-
-	return utils.NewGenericBucket(buckets), currBucket, nil
-}
-
 func adjustForBaseUsage(newCPUPerReplica, currentperReplicaCPU, newMemPerReplica, currentperReplicaMem int64, finalReplica, currentReplicas int32, hvpa *hvpav1alpha1.Hvpa) (newCPU, newMem int64) {
 	baseUsageSpec := hvpa.Spec.BaseResourcesPerReplica
 	if baseUsageSpec == nil || finalReplica == currentReplicas {
@@ -275,7 +132,7 @@ func getScalingRecommendations(
 	}
 
 	limitScalingParams := hvpa.Spec.Vpa.LimitsRequestsGapScaleParams
-	buckets, currentBucket, err := getBuckets(hvpa, currentReplicas)
+	buckets, currentBucket, err := GetBuckets(hvpa, currentReplicas)
 	if err != nil {
 		return nil, nil, false, nil, err
 	}
